backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
sqlalchemy==2.0.31
asyncpg==0.29.0
python-dotenv==1.0.1
